refactor: change photos on hero-section and features-section
