+++ src/components/hero-section/hero_section.scss
@@ -1,20 +1,22 @@
 @use '@/scss/base/variables' as var;
-@use '@/components/ui/header';
+@use '@/components/header/header.scss';
 
 .hero {
   display: flex;
   width: 100%;
   height: 112vh;
-  padding-top: 20vw;
+  padding-top: 15vw;
   padding-left: 3vw;
-  padding-right: 3vw;
+  padding-right: 5vw;
+  gap: 10vw;
 
   &-content {
-    width: 60%;
+    width: 50%;
     height: 45vh;
     display: flex;
     flex-direction: column;
     justify-content: center;
+    margin-top: 8vh;
     gap: 1vw;
 
     &__headline {
@@ -37,11 +39,13 @@
   }
 
   &-image {
-    width: 40%;
+    width: 50%;
+    justify-content: center;
+    align-items: center;
 
     &__image {
       width: 100%;
-      height: auto;
+      height: 65vh;
       object-fit: cover;
       border-radius: var.$radius-lg;
     }
