--- src/components/ui/capabilities_section.scss
@@ -1,73 +1,0 @@
-@use '@/scss/base/variables' as var;
-@use '@/components/ui/header';
-
-.capabilities {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  padding-top: 20vh;
-  padding-bottom: 20vh;
-  border-bottom: #e5e7eb 1px solid;
-
-  &__heading {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    gap: 10px;
-  }
-
-  &__title {
-  }
-
-  &__subtitle {
-    font-size: 1.3rem;
-    color: dimgrey;
-  }
-
-  &__container {
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    align-items: center;
-    gap: 30px;
-    margin-top: 7vh;
-  }
-}
-
-.capability-card {
-  width: auto;
-  padding: 24px;
-  border: #e5e7eb 1px solid;
-  box-shadow: var.$shadow;
-  border-radius: var.$radius-lg;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  text-align: center;
-
-  &__icon{
-    width: 45px;
-    height: 45px;
-    border-radius: 50%;
-    background-color: lightgray;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-  }
-
-  &__title{
-    margin-top: 8px;
-  }
-
-  &__text{
-    font-size: var.$text-sm;
-    color: dimgrey;
-    text-wrap: wrap;
-    width: 260px;
-    line-height: var.$leading-tight;
-    margin: 8px;
-  }
-}
