+++ src/pages/home/home.tsx
@@ -1,9 +1,9 @@
-import { Header } from '@/components/header.tsx';
-import { HeroSection } from '@/components/hero_section.tsx';
-import { FeaturesSection } from '@/components/features_section.tsx';
-import { CapabilitiesSection } from '@/components/capabilities-section.tsx';
-import { ContactSection } from '@/components/contact-section.tsx';
-import { Footer } from '@/components/footer.tsx';
+import { Header } from '@/components/header/header.tsx';
+import { HeroSection } from '@/components/hero-section/hero_section.tsx';
+import { FeaturesSection } from '@/components/features-section/features_section.tsx';
+import { CapabilitiesSection } from '@/components/capabilities-section/capabilities-section.tsx';
+import { ContactSection } from '@/components/contact-section/contact-section.tsx';
+import { Footer } from '@/components/footer/footer.tsx';
 
 export const Home = () => {
   return (
